Update README.md
This change adds a contact email for the CulturePlay Learning and Quiz Program to the README file. The email address is included to provide users with a way to ask questions and provide feedback on the program. By including this contact information, we hope to make the program more accessible and user-friendly.